Deduplicate error display code
